MetricsReporterOptions goes into separate file.
